Handle case where user sends GC as gift via PPDG
When a user sends a gift card as a gift instead of buying for one's self, PPDG renders some additional content before rendering the skip button. As such, in its current state, the code is evaluated before there's a skip button to click. This amendment attaches a DOM observer and will attempt the skip button logic for each time a dom element is added or removed. Attribute changes are ignored. An initial attempt at clicking the skip button is still made for pages where this is not an issue.